Refactor - extract common activity code to superclass
